upd: make.sh

+++ make.sh
@@ -5,6 +5,7 @@ SRC_PATH="src/"
 RESOURCES="main/resources/" # must be in the src/ folder
 OUT_PATH="out/"
 DOC_PATH="doc/"
+TEST_PATH="./test*"
 
 # Commands
 BUILD="build"
@@ -21,19 +22,8 @@ HELP="help"
 HELP_DESC="Display the help menu"
 
 # Files list
-declare -a CLASSES=(
-  "main/java/fr/alexandreladriere/shortestpath/ShortestPath.java"
-  "main/java/fr/alexandreladriere/shortestpath/utils/Colors.java"
-  "main/java/fr/alexandreladriere/shortestpath/utils/Constants.java"
-  "main/java/fr/alexandreladriere/shortestpath/utils/Strings.java"
-  "main/java/fr/alexandreladriere/shortestpath/models/bfs/BreadthFirstSearch.java"
-  "main/java/fr/alexandreladriere/shortestpath/models/bfs/Node.java"
-  "main/java/fr/alexandreladriere/shortestpath/gui/Case.java"
-  "main/java/fr/alexandreladriere/shortestpath/gui/Controller.java"
-  "main/java/fr/alexandreladriere/shortestpath/gui/MouseController.java"
-  "main/java/fr/alexandreladriere/shortestpath/gui/Popup.java"
-  "main/java/fr/alexandreladriere/shortestpath/gui/Gui.java"
-)
+declare -a CLASSES=()
+declare -a EXCLUDE=()
 
 # Main class name (without extension)
 MAIN="main/java/fr/alexandreladriere/shortestpath/ShortestPath"
@@ -43,6 +33,7 @@ CLASS_STR=""
 function build() {
   # Builds all .class files which are in the $CLASSES array into the $OUT_PATH folder
   echo "Building .class files..."
+  finding_files # Fill the .java files list
   mkdir -p $OUT_PATH
 	cd $SRC_PATH || exit
 	echo "cd $SRC_PATH"
@@ -80,8 +71,10 @@ function make_class_string() {
   # transforms the list of all java files to a single string with .class extension for each file)
   for file in "${CLASSES[@]}"
   do
-    tmp="${file%.*}.class" # Replace ".java" by ".class"
-    CLASS_STR="${CLASS_STR} $tmp"
+    if [ "$file" != "" ]; then
+      tmp="${file%.*}.class" # Replace ".java" by ".class"
+      CLASS_STR="${CLASS_STR} $tmp"
+    fi
   done
 }
 
@@ -95,6 +88,7 @@ function build_jar() {
     echo "${MAIN}.class does not exist"
     build
   fi
+  finding_files
   cd $OUT_PATH || exit
   echo "cd $OUT_PATH"
   make_class_string
@@ -109,20 +103,18 @@ function build_jar() {
 function clean() {
   # Removes the $OUT_PATH folder
   echo "Cleaning folder..."
-  if [ ! -d "$OUT_PATH" ]; then # checks if $OUT_PATH exists]
-    echo "Nothing to clean"
-  else
-    rm -r $OUT_PATH
-    rm -r $DOC_PATH
-    echo "Folder cleaned"
-    exit
-  fi
+  [ -e $OUT_PATH ] && rm -r $OUT_PATH
+  [ -e $DOC_PATH ] && rm -r $DOC_PATH
+  [ -e ${MAIN##*/}.jar ] && rm ${MAIN##*/}.jar # remove .jar file
+  echo "Folder cleaned"
   exit
+  e
 }
 
 function create_javadoc() {
   # Create all javadoc
   echo "Creating javadoc..."
+  finding_files
   mkdir -p $DOC_PATH
   main_package=${MAIN////.}
   echo "javadoc -d ${DOC_PATH} -sourcepath $SRC_PATH -subpackages ${main_package%.*}"
@@ -142,6 +134,22 @@ function help() {
   echo "$CLEAN         ->     $CLEAN_DESC"
   echo "$HELP          ->     $HELP_DESC"
   exit
+}
+
+function finding_files() {
+  # fill the .java file list used by other functions
+  cd $SRC_PATH || exit
+  CLASSES=($(find ./ -not -path "$TEST_PATH" -type f))
+  remove_exclude_files
+  cd ..
+}
+
+function remove_exclude_files() {
+  # remove all .java files that you want to exclude from the CLASSES array
+  for exclude in "${EXCLUDE[@]}"
+  do
+    CLASSES=("${CLASSES[@]/$exclude}")
+  done
 }
 
 ### Main body ###
